Update pos or neg or zero.py

+++ Beginner/pos or neg or zero.py
@@ -1,8 +1,8 @@
 a= input("enter the input:")
 try:
-    if (int(a)>0):
+    if (float(a)>0):
         print("The given number is positive")
-    elif(int(a)==0):
+    elif(float(a)==0):
         print("The given interger is ze")
     else :
         print("The given interger is ne")
